Able to generate player input based on user selection, unable to send all player info to DB
